+++ src/CountryInfo.php
@@ -17,10 +17,8 @@ use Iban\Validation\Swift\Registry;
  * Represents IBAN country information.
  *
  * @author Jan Schädlich <mail@janschaedlich.de>
- *
- * @final since 1.7
  */
-class CountryInfo
+final class CountryInfo
 {
     /**
      * @var string
@@ -75,22 +73,6 @@ class CountryInfo
     public function getBbanLength(): int
     {
         return $this->swiftRegistry->getBbanLength($this->countryCode);
-    }
-
-    /**
-     * @deprecated since 1.7
-     */
-    public function getBbanBankIdentifierStartPos(): int
-    {
-        return $this->swiftRegistry->getBbanBankIdentifierStartPos($this->countryCode);
-    }
-
-    /**
-     * @deprecated since 1.7
-     */
-    public function getBbanBankIdentifierEndPos(): int
-    {
-        return $this->swiftRegistry->getBbanBankIdentifierEndPos($this->countryCode);
     }
 
     public function getIbanElectronicExample(): string